start breaking down the plotting file
